moved config drop component to separate directory and added initial snapshot test
